Move transaction output related code to `Cardano.Api.Internal.Tx.Output`
