+++ cardano-api/cardano-api.cabal
@@ -268,6 +268,7 @@ library
     Cardano.Api.Internal.SpecialByron
     Cardano.Api.Internal.StakePoolMetadata
     Cardano.Api.Internal.Tx.BuildTxWith
+    Cardano.Api.Internal.Tx.Output
     Cardano.Api.Internal.Tx.UTxO
     Cardano.Api.Internal.TxIn
     Cardano.Api.Internal.TxMetadata
