+++ cardano-api/src/Cardano/Api/Internal/Tx/Body.hs
@@ -1,7 +1,6 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DerivingVia #-}
 {-# LANGUAGE DisambiguateRecordFields #-}
-{-# LANGUAGE EmptyCase #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GADTs #-}
@@ -260,6 +259,7 @@ module Cardano.Api.Internal.Tx.Body
   , setTxCurrentTreasuryValue
   , setTxTreasuryDonation
   , TxBodyError (..)
+  , TxOutputError (..)
   , TxBodyScriptData (..)
   , TxScriptValidity (..)
   , ScriptValidity (..)
@@ -408,7 +408,7 @@ import Cardano.Api.Internal.Eon.ShelleyBasedEra
 import Cardano.Api.Internal.Eon.ShelleyToBabbageEra
 import Cardano.Api.Internal.Eras.Case
 import Cardano.Api.Internal.Eras.Core
-import Cardano.Api.Internal.Error (Error (..), displayError)
+import Cardano.Api.Internal.Error
 import Cardano.Api.Internal.Experimental.Plutus.IndexedPlutusScriptWitness
   ( Witnessable (..)
   , WitnessableItem (..)
@@ -428,16 +428,14 @@ import Cardano.Api.Internal.ProtocolParameters
 import Cardano.Api.Internal.ReexposeLedger qualified as Ledger
 import Cardano.Api.Internal.Script
 import Cardano.Api.Internal.ScriptData
-import Cardano.Api.Internal.Serialise.Cbor
 import Cardano.Api.Internal.SerialiseJSON
-import Cardano.Api.Internal.SerialiseRaw
 import Cardano.Api.Internal.Tx.BuildTxWith
+import Cardano.Api.Internal.Tx.Output
 import Cardano.Api.Internal.Tx.Sign
 import Cardano.Api.Internal.TxIn
 import Cardano.Api.Internal.TxMetadata
 import Cardano.Api.Internal.Utils
 import Cardano.Api.Internal.Value
-import Cardano.Api.Internal.ValueParser
 import Cardano.Api.Ledger.Lens qualified as A
 
 import Cardano.Chain.Common qualified as Byron
@@ -457,13 +455,11 @@ import Cardano.Ledger.Binary qualified as CBOR
 import Cardano.Ledger.Coin qualified as L
 import Cardano.Ledger.Conway.Core qualified as L
 import Cardano.Ledger.Core ()
-import Cardano.Ledger.Core qualified as Core
 import Cardano.Ledger.Core qualified as Ledger
 import Cardano.Ledger.Credential qualified as Shelley
 import Cardano.Ledger.Hashes qualified as SafeHash
 import Cardano.Ledger.Keys qualified as Shelley
 import Cardano.Ledger.Mary.Value as L (MaryValue (..), MultiAsset)
-import Cardano.Ledger.Plutus.Data qualified as Plutus
 import Cardano.Ledger.Plutus.Language qualified as Plutus
 import Cardano.Ledger.Shelley.API qualified as Ledger
 import Cardano.Ledger.Shelley.Genesis qualified as Shelley
@@ -480,16 +476,11 @@ import Ouroboros.Consensus.Shelley.Eras qualified as E
   , ShelleyEra
   )
 
-import Control.Applicative
 import Control.Monad
-import Data.Aeson (object, withObject, (.:), (.:?), (.=))
+import Data.Aeson (object, (.=))
 import Data.Aeson qualified as Aeson
-import Data.Aeson.Key qualified as Aeson
-import Data.Aeson.Types qualified as Aeson
 import Data.Bifunctor (Bifunctor (..))
 import Data.ByteString (ByteString)
-import Data.ByteString.Base16 qualified as Base16
-import Data.ByteString.Char8 qualified as BSC
 import Data.Foldable (for_)
 import Data.Function (on)
 import Data.Functor (($>))
@@ -504,14 +495,10 @@ import Data.MonoTraversable (omap)
 import Data.Monoid
 import Data.OSet.Strict (OSet)
 import Data.OSet.Strict qualified as OSet
-import Data.Scientific (toBoundedInteger)
 import Data.Sequence.Strict qualified as Seq
 import Data.Set (Set)
 import Data.Set qualified as Set
-import Data.String
 import Data.Text (Text)
-import Data.Text qualified as Text
-import Data.Text.Encoding qualified as Text
 import Data.Text.Lazy qualified as LText
 import Data.Text.Lazy.Builder qualified as LText
 import Data.Type.Equality
@@ -523,589 +510,6 @@ import GHC.Exts (IsList (..))
 import GHC.Stack
 import Lens.Micro hiding (ix)
 import Lens.Micro.Extras (view)
-import Text.Parsec ((<?>))
-import Text.Parsec qualified as Parsec
-import Text.Parsec.String qualified as Parsec
-
--- ----------------------------------------------------------------------------
--- Transaction outputs
---
-
--- | The context is a transaction body
-data CtxTx
-
--- | The context is the UTxO
-data CtxUTxO
-
-data TxOut ctx era
-  = TxOut
-      (AddressInEra era)
-      (TxOutValue era)
-      (TxOutDatum ctx era)
-      (ReferenceScript era)
-
-deriving instance Eq (TxOut ctx era)
-
-deriving instance Show (TxOut ctx era)
-
-data TxOutInAnyEra where
-  TxOutInAnyEra
-    :: CardanoEra era
-    -> TxOut CtxTx era
-    -> TxOutInAnyEra
-
-deriving instance Show TxOutInAnyEra
-
-instance Eq TxOutInAnyEra where
-  TxOutInAnyEra era1 out1 == TxOutInAnyEra era2 out2 =
-    case testEquality era1 era2 of
-      Just Refl -> out1 == out2
-      Nothing -> False
-
-deriving via (ShowOf TxOutInAnyEra) instance Pretty TxOutInAnyEra
-
--- | Convenience constructor for 'TxOutInAnyEra'
-txOutInAnyEra :: CardanoEra era -> TxOut CtxTx era -> TxOutInAnyEra
-txOutInAnyEra = TxOutInAnyEra
-
-toCtxUTxOTxOut :: TxOut CtxTx era -> TxOut CtxUTxO era
-toCtxUTxOTxOut (TxOut addr val d refS) =
-  let dat = case d of
-        TxOutDatumNone -> TxOutDatumNone
-        TxOutDatumHash s h -> TxOutDatumHash s h
-        TxOutSupplementalDatum s datum -> TxOutDatumHash s $ hashScriptDataBytes datum
-        TxOutDatumInline s sd -> TxOutDatumInline s sd
-   in TxOut addr val dat refS
-
-fromCtxUTxOTxOut :: TxOut CtxUTxO era -> TxOut CtxTx era
-fromCtxUTxOTxOut (TxOut addr val d refS) =
-  let dat = case d of
-        TxOutDatumNone -> TxOutDatumNone
-        TxOutDatumHash s h -> TxOutDatumHash s h
-        TxOutDatumInline s sd -> TxOutDatumInline s sd
-   in TxOut addr val dat refS
-
-instance IsCardanoEra era => ToJSON (TxOut ctx era) where
-  toJSON = txOutToJsonValue cardanoEra
-
-txOutToJsonValue :: CardanoEra era -> TxOut ctx era -> Aeson.Value
-txOutToJsonValue era (TxOut addr val dat refScript) =
-  case era of
-    ByronEra -> object ["address" .= addr, "value" .= val]
-    ShelleyEra -> object ["address" .= addr, "value" .= val]
-    AllegraEra -> object ["address" .= addr, "value" .= val]
-    MaryEra -> object ["address" .= addr, "value" .= val]
-    AlonzoEra ->
-      object
-        [ "address" .= addr
-        , "value" .= val
-        , datHashJsonVal dat
-        , "datum" .= datJsonVal dat
-        ]
-    BabbageEra ->
-      object
-        [ "address" .= addr
-        , "value" .= val
-        , datHashJsonVal dat
-        , "datum" .= datJsonVal dat
-        , "inlineDatum" .= inlineDatumJsonVal dat
-        , "inlineDatumRaw" .= inlineDatumRawJsonCbor dat
-        , "referenceScript" .= refScriptJsonVal refScript
-        ]
-    ConwayEra ->
-      object
-        [ "address" .= addr
-        , "value" .= val
-        , datHashJsonVal dat
-        , "datum" .= datJsonVal dat
-        , "inlineDatum" .= inlineDatumJsonVal dat
-        , "inlineDatumRaw" .= inlineDatumRawJsonCbor dat
-        , "referenceScript" .= refScriptJsonVal refScript
-        ]
- where
-  datHashJsonVal :: TxOutDatum ctx era -> Aeson.Pair
-  datHashJsonVal d =
-    case d of
-      TxOutDatumNone ->
-        "datumhash" .= Aeson.Null
-      TxOutDatumHash _ h ->
-        "datumhash" .= toJSON h
-      TxOutSupplementalDatum _ datum ->
-        "datumhash" .= toJSON (hashScriptDataBytes datum)
-      TxOutDatumInline _ datum ->
-        "inlineDatumhash" .= toJSON (hashScriptDataBytes datum)
-
-  datJsonVal :: TxOutDatum ctx era -> Aeson.Value
-  datJsonVal d =
-    case d of
-      TxOutDatumNone -> Aeson.Null
-      TxOutDatumHash _ _ -> Aeson.Null
-      TxOutSupplementalDatum _ datum -> scriptDataToJson ScriptDataJsonDetailedSchema datum
-      TxOutDatumInline _ _ -> Aeson.Null
-
-  inlineDatumJsonVal :: TxOutDatum ctx era -> Aeson.Value
-  inlineDatumJsonVal d =
-    case d of
-      TxOutDatumNone -> Aeson.Null
-      TxOutDatumHash{} -> Aeson.Null
-      TxOutSupplementalDatum{} -> Aeson.Null
-      TxOutDatumInline _ datum -> scriptDataToJson ScriptDataJsonDetailedSchema datum
-
-  inlineDatumRawJsonCbor :: TxOutDatum ctx era -> Aeson.Value
-  inlineDatumRawJsonCbor d =
-    case d of
-      TxOutDatumNone -> Aeson.Null
-      TxOutDatumHash{} -> Aeson.Null
-      TxOutSupplementalDatum{} -> Aeson.Null
-      TxOutDatumInline _ datum ->
-        Aeson.String
-          . Text.decodeUtf8
-          . Base16.encode
-          . serialiseToCBOR
-          $ datum
-
-  refScriptJsonVal :: ReferenceScript era -> Aeson.Value
-  refScriptJsonVal rScript =
-    case rScript of
-      ReferenceScript _ s -> toJSON s
-      ReferenceScriptNone -> Aeson.Null
-
-instance IsShelleyBasedEra era => FromJSON (TxOut CtxTx era) where
-  parseJSON = withObject "TxOut" $ \o -> do
-    case shelleyBasedEra :: ShelleyBasedEra era of
-      ShelleyBasedEraShelley ->
-        TxOut
-          <$> o .: "address"
-          <*> o .: "value"
-          <*> return TxOutDatumNone
-          <*> return ReferenceScriptNone
-      ShelleyBasedEraMary ->
-        TxOut
-          <$> o .: "address"
-          <*> o .: "value"
-          <*> return TxOutDatumNone
-          <*> return ReferenceScriptNone
-      ShelleyBasedEraAllegra ->
-        TxOut
-          <$> o .: "address"
-          <*> o .: "value"
-          <*> return TxOutDatumNone
-          <*> return ReferenceScriptNone
-      ShelleyBasedEraAlonzo -> alonzoTxOutParser AlonzoEraOnwardsAlonzo o
-      ShelleyBasedEraBabbage -> do
-        alonzoTxOutInBabbage <- alonzoTxOutParser AlonzoEraOnwardsBabbage o
-
-        -- We check for the existence of inline datums
-        inlineDatumHash <- o .:? "inlineDatumhash"
-        inlineDatum <- o .:? "inlineDatum"
-        mInlineDatum <-
-          case (inlineDatum, inlineDatumHash) of
-            (Just dVal, Just h) -> do
-              case scriptDataJsonToHashable ScriptDataJsonDetailedSchema dVal of
-                Left err ->
-                  fail $ "Error parsing TxOut JSON: " <> displayError err
-                Right hashableData -> do
-                  if hashScriptDataBytes hashableData /= h
-                    then fail "Inline datum not equivalent to inline datum hash"
-                    else return $ TxOutDatumInline BabbageEraOnwardsBabbage hashableData
-            (Nothing, Nothing) -> return TxOutDatumNone
-            (_, _) ->
-              fail
-                "Should not be possible to create a tx output with either an inline datum hash or an inline datum"
-
-        mReferenceScript <- o .:? "referenceScript"
-
-        reconcileBabbage alonzoTxOutInBabbage mInlineDatum mReferenceScript
-      ShelleyBasedEraConway -> do
-        alonzoTxOutInConway <- alonzoTxOutParser AlonzoEraOnwardsConway o
-
-        -- We check for the existence of inline datums
-        inlineDatumHash <- o .:? "inlineDatumhash"
-        inlineDatum <- o .:? "inlineDatum"
-        mInlineDatum <-
-          case (inlineDatum, inlineDatumHash) of
-            (Just dVal, Just h) ->
-              case scriptDataFromJson ScriptDataJsonDetailedSchema dVal of
-                Left err ->
-                  fail $ "Error parsing TxOut JSON: " <> displayError err
-                Right sData ->
-                  if hashScriptDataBytes sData /= h
-                    then fail "Inline datum not equivalent to inline datum hash"
-                    else return $ TxOutDatumInline BabbageEraOnwardsConway sData
-            (Nothing, Nothing) -> return TxOutDatumNone
-            (_, _) ->
-              fail
-                "Should not be possible to create a tx output with either an inline datum hash or an inline datum"
-
-        mReferenceScript <- o .:? "referenceScript"
-
-        reconcileConway alonzoTxOutInConway mInlineDatum mReferenceScript
-   where
-    reconcileBabbage
-      :: TxOut CtxTx BabbageEra
-      -- \^ Alonzo era datum in Babbage era
-      -> TxOutDatum CtxTx BabbageEra
-      -- \^ Babbage inline datum
-      -> Maybe ScriptInAnyLang
-      -> Aeson.Parser (TxOut CtxTx BabbageEra)
-    reconcileBabbage top@(TxOut addr v dat r) babbageDatum mBabRefScript = do
-      -- We check for conflicting datums
-      finalDat <- case (dat, babbageDatum) of
-        (TxOutDatumNone, bDatum) -> return bDatum
-        (anyDat, TxOutDatumNone) -> return anyDat
-        (alonzoDat, babbageDat) ->
-          fail $
-            "Parsed an Alonzo era datum and a Babbage era datum "
-              <> "TxOut: "
-              <> show top
-              <> "Alonzo datum: "
-              <> show alonzoDat
-              <> "Babbage dat: "
-              <> show babbageDat
-      finalRefScript <- case mBabRefScript of
-        Nothing -> return r
-        Just anyScript ->
-          return $ ReferenceScript BabbageEraOnwardsBabbage anyScript
-      return $ TxOut addr v finalDat finalRefScript
-
-    reconcileConway
-      :: TxOut CtxTx ConwayEra
-      -- \^ Alonzo era datum in Conway era
-      -> TxOutDatum CtxTx ConwayEra
-      -- \^ Babbage inline datum
-      -> Maybe ScriptInAnyLang
-      -> Aeson.Parser (TxOut CtxTx ConwayEra)
-    reconcileConway top@(TxOut addr v dat r) babbageDatum mBabRefScript = do
-      -- We check for conflicting datums
-      finalDat <- case (dat, babbageDatum) of
-        (TxOutDatumNone, bDatum) -> return bDatum
-        (anyDat, TxOutDatumNone) -> return anyDat
-        (alonzoDat, babbageDat) ->
-          fail $
-            "Parsed an Alonzo era datum and a Conway era datum "
-              <> "TxOut: "
-              <> show top
-              <> "Alonzo datum: "
-              <> show alonzoDat
-              <> "Conway dat: "
-              <> show babbageDat
-      finalRefScript <- case mBabRefScript of
-        Nothing -> return r
-        Just anyScript ->
-          return $ ReferenceScript BabbageEraOnwardsConway anyScript
-      return $ TxOut addr v finalDat finalRefScript
-
-    alonzoTxOutParser
-      :: AlonzoEraOnwards era -> Aeson.Object -> Aeson.Parser (TxOut CtxTx era)
-    alonzoTxOutParser w o = do
-      mDatumHash <- o .:? "datumhash"
-      mDatumVal <- o .:? "datum"
-      case (mDatumVal, mDatumHash) of
-        (Nothing, Nothing) ->
-          TxOut
-            <$> o .: "address"
-            <*> o .: "value"
-            <*> return TxOutDatumNone
-            <*> return ReferenceScriptNone
-        (Just dVal, Just{}) -> do
-          case scriptDataJsonToHashable ScriptDataJsonDetailedSchema dVal of
-            Left e -> fail $ "Error parsing ScriptData: " <> show e
-            Right hashableData ->
-              TxOut
-                <$> o .: "address"
-                <*> o .: "value"
-                <*> return (TxOutSupplementalDatum w hashableData)
-                <*> return ReferenceScriptNone
-        (Nothing, Just dHash) ->
-          TxOut
-            <$> o .: "address"
-            <*> o .: "value"
-            <*> return (TxOutDatumHash w dHash)
-            <*> return ReferenceScriptNone
-        (Just _dVal, Nothing) -> fail "Only datum JSON was found, this should not be possible."
-
-instance IsShelleyBasedEra era => FromJSON (TxOut CtxUTxO era) where
-  parseJSON = withObject "TxOut" $ \o -> do
-    case shelleyBasedEra :: ShelleyBasedEra era of
-      ShelleyBasedEraShelley ->
-        TxOut
-          <$> o .: "address"
-          <*> o .: "value"
-          <*> return TxOutDatumNone
-          <*> return ReferenceScriptNone
-      ShelleyBasedEraMary ->
-        TxOut
-          <$> o .: "address"
-          <*> o .: "value"
-          <*> return TxOutDatumNone
-          <*> return ReferenceScriptNone
-      ShelleyBasedEraAllegra ->
-        TxOut
-          <$> o .: "address"
-          <*> o .: "value"
-          <*> return TxOutDatumNone
-          <*> return ReferenceScriptNone
-      ShelleyBasedEraAlonzo -> alonzoTxOutParser AlonzoEraOnwardsAlonzo o
-      ShelleyBasedEraBabbage -> do
-        alonzoTxOutInBabbage <- alonzoTxOutParser AlonzoEraOnwardsBabbage o
-
-        -- We check for the existence of inline datums
-        inlineDatumHash <- o .:? "inlineDatumhash"
-        inlineDatum <- o .:? "inlineDatum"
-        mInlineDatum <-
-          case (inlineDatum, inlineDatumHash) of
-            (Just dVal, Just h) -> do
-              case scriptDataJsonToHashable ScriptDataJsonDetailedSchema dVal of
-                Left err ->
-                  fail $ "Error parsing TxOut JSON: " <> displayError err
-                Right hashableData -> do
-                  if hashScriptDataBytes hashableData /= h
-                    then fail "Inline datum not equivalent to inline datum hash"
-                    else return $ TxOutDatumInline BabbageEraOnwardsBabbage hashableData
-            (Nothing, Nothing) -> return TxOutDatumNone
-            (_, _) ->
-              fail
-                "Should not be possible to create a tx output with either an inline datum hash or an inline datum"
-
-        -- We check for a reference script
-        mReferenceScript <- o .:? "referenceScript"
-
-        reconcileBabbage alonzoTxOutInBabbage mInlineDatum mReferenceScript
-      ShelleyBasedEraConway -> do
-        alonzoTxOutInConway <- alonzoTxOutParser AlonzoEraOnwardsConway o
-
-        -- We check for the existence of inline datums
-        inlineDatumHash <- o .:? "inlineDatumhash"
-        inlineDatum <- o .:? "inlineDatum"
-        mInlineDatum <-
-          case (inlineDatum, inlineDatumHash) of
-            (Just dVal, Just h) ->
-              case scriptDataFromJson ScriptDataJsonDetailedSchema dVal of
-                Left err ->
-                  fail $ "Error parsing TxOut JSON: " <> displayError err
-                Right sData ->
-                  if hashScriptDataBytes sData /= h
-                    then fail "Inline datum not equivalent to inline datum hash"
-                    else return $ TxOutDatumInline BabbageEraOnwardsConway sData
-            (Nothing, Nothing) -> return TxOutDatumNone
-            (_, _) ->
-              fail
-                "Should not be possible to create a tx output with either an inline datum hash or an inline datum"
-
-        -- We check for a reference script
-        mReferenceScript <- o .:? "referenceScript"
-
-        reconcileConway alonzoTxOutInConway mInlineDatum mReferenceScript
-   where
-    reconcileBabbage
-      :: TxOut CtxUTxO BabbageEra
-      -- \^ Alonzo era datum in Babbage era
-      -> TxOutDatum CtxUTxO BabbageEra
-      -- \^ Babbage inline datum
-      -> Maybe ScriptInAnyLang
-      -> Aeson.Parser (TxOut CtxUTxO BabbageEra)
-    reconcileBabbage (TxOut addr v dat r) babbageDatum mBabRefScript = do
-      -- We check for conflicting datums
-      finalDat <- case (dat, babbageDatum) of
-        (TxOutDatumNone, bDatum) -> return bDatum
-        (anyDat, TxOutDatumNone) -> return anyDat
-        (_, _) -> fail "Parsed an Alonzo era datum and a Babbage era datum"
-      finalRefScript <- case mBabRefScript of
-        Nothing -> return r
-        Just anyScript ->
-          return $ ReferenceScript BabbageEraOnwardsBabbage anyScript
-
-      return $ TxOut addr v finalDat finalRefScript
-
-    reconcileConway
-      :: TxOut CtxUTxO ConwayEra
-      -- \^ Alonzo era datum in Conway era
-      -> TxOutDatum CtxUTxO ConwayEra
-      -- \^ Babbage inline datum
-      -> Maybe ScriptInAnyLang
-      -> Aeson.Parser (TxOut CtxUTxO ConwayEra)
-    reconcileConway (TxOut addr v dat r) babbageDatum mBabRefScript = do
-      -- We check for conflicting datums
-      finalDat <- case (dat, babbageDatum) of
-        (TxOutDatumNone, bDatum) -> return bDatum
-        (anyDat, TxOutDatumNone) -> return anyDat
-        (_, _) -> fail "Parsed an Alonzo era datum and a Conway era datum"
-      finalRefScript <- case mBabRefScript of
-        Nothing -> return r
-        Just anyScript ->
-          return $ ReferenceScript BabbageEraOnwardsConway anyScript
-
-      return $ TxOut addr v finalDat finalRefScript
-
-    alonzoTxOutParser :: AlonzoEraOnwards era -> Aeson.Object -> Aeson.Parser (TxOut CtxUTxO era)
-    alonzoTxOutParser w o = do
-      mDatumHash <- o .:? "datumhash"
-      case mDatumHash of
-        Nothing ->
-          TxOut
-            <$> o .: "address"
-            <*> o .: "value"
-            <*> return TxOutDatumNone
-            <*> return ReferenceScriptNone
-        Just dHash ->
-          TxOut
-            <$> o .: "address"
-            <*> o .: "value"
-            <*> return (TxOutDatumHash w dHash)
-            <*> return ReferenceScriptNone
-
-toByronTxOut :: TxOut ctx ByronEra -> Maybe Byron.TxOut
-toByronTxOut = \case
-  TxOut (AddressInEra ByronAddressInAnyEra (ByronAddress addr)) (TxOutValueByron value) _ _ ->
-    Byron.TxOut addr <$> toByronLovelace value
-  TxOut (AddressInEra ByronAddressInAnyEra (ByronAddress _)) (TxOutValueShelleyBased w _) _ _ ->
-    case w of {}
-  TxOut (AddressInEra (ShelleyAddressInEra sbe) ShelleyAddress{}) _ _ _ ->
-    case sbe of {}
-
-toShelleyTxOut
-  :: forall era ledgerera
-   . HasCallStack
-  => ShelleyLedgerEra era ~ ledgerera
-  => ShelleyBasedEra era
-  -> TxOut CtxUTxO era
-  -> Ledger.TxOut ledgerera
-toShelleyTxOut sbe = shelleyBasedEraConstraints sbe $ \case
-  TxOut addr (TxOutValueShelleyBased _ value) txoutdata refScript ->
-    caseShelleyToMaryOrAlonzoEraOnwards
-      (const $ L.mkBasicTxOut (toShelleyAddr addr) value)
-      ( \case
-          AlonzoEraOnwardsAlonzo ->
-            L.mkBasicTxOut (toShelleyAddr addr) value
-              & L.dataHashTxOutL
-                .~ toAlonzoTxOutDatumHashUTxO txoutdata
-          AlonzoEraOnwardsBabbage ->
-            L.mkBasicTxOut (toShelleyAddr addr) value
-              & L.datumTxOutL
-                .~ toBabbageTxOutDatum txoutdata
-              & L.referenceScriptTxOutL
-                .~ refScriptToShelleyScript sbe refScript
-          AlonzoEraOnwardsConway ->
-            L.mkBasicTxOut (toShelleyAddr addr) value
-              & L.datumTxOutL
-                .~ toBabbageTxOutDatumUTxO txoutdata
-              & L.referenceScriptTxOutL
-                .~ refScriptToShelleyScript sbe refScript
-      )
-      sbe
-
-toAlonzoTxOutDatumHashUTxO
-  :: TxOutDatum CtxUTxO era -> StrictMaybe Plutus.DataHash
-toAlonzoTxOutDatumHashUTxO TxOutDatumNone = SNothing
-toAlonzoTxOutDatumHashUTxO (TxOutDatumHash _ (ScriptDataHash dh)) = SJust dh
-toAlonzoTxOutDatumHashUTxO (TxOutDatumInline{}) = SNothing
-
-toBabbageTxOutDatumUTxO
-  :: L.Era (ShelleyLedgerEra era)
-  => TxOutDatum CtxUTxO era
-  -> Plutus.Datum (ShelleyLedgerEra era)
-toBabbageTxOutDatumUTxO TxOutDatumNone = Plutus.NoDatum
-toBabbageTxOutDatumUTxO (TxOutDatumHash _ (ScriptDataHash dh)) = Plutus.DatumHash dh
-toBabbageTxOutDatumUTxO (TxOutDatumInline _ sd) = scriptDataToInlineDatum sd
-
-fromShelleyTxOut
-  :: forall era ctx
-   . ()
-  => ShelleyBasedEra era
-  -> Core.TxOut (ShelleyLedgerEra era)
-  -> TxOut ctx era
-fromShelleyTxOut sbe ledgerTxOut = shelleyBasedEraConstraints sbe $ do
-  let txOutValue = TxOutValueShelleyBased sbe $ ledgerTxOut ^. A.valueTxOutL sbe
-  let addressInEra = fromShelleyAddr sbe $ ledgerTxOut ^. L.addrTxOutL
-
-  case sbe of
-    ShelleyBasedEraShelley ->
-      TxOut addressInEra txOutValue TxOutDatumNone ReferenceScriptNone
-    ShelleyBasedEraAllegra ->
-      TxOut addressInEra txOutValue TxOutDatumNone ReferenceScriptNone
-    ShelleyBasedEraMary ->
-      TxOut addressInEra txOutValue TxOutDatumNone ReferenceScriptNone
-    ShelleyBasedEraAlonzo ->
-      TxOut
-        addressInEra
-        txOutValue
-        (fromAlonzoTxOutDatumHash AlonzoEraOnwardsAlonzo mDatumHash)
-        ReferenceScriptNone
-     where
-      mDatumHash = ledgerTxOut ^. L.dataHashTxOutL
-    ShelleyBasedEraBabbage ->
-      TxOut
-        addressInEra
-        txOutValue
-        ( fromBabbageTxOutDatum
-            AlonzoEraOnwardsBabbage
-            BabbageEraOnwardsBabbage
-            datum
-        )
-        ( case mRefScript of
-            SNothing -> ReferenceScriptNone
-            SJust refScript ->
-              fromShelleyScriptToReferenceScript ShelleyBasedEraBabbage refScript
-        )
-     where
-      datum = ledgerTxOut ^. L.datumTxOutL
-      mRefScript = ledgerTxOut ^. L.referenceScriptTxOutL
-    ShelleyBasedEraConway ->
-      TxOut
-        addressInEra
-        txOutValue
-        ( fromBabbageTxOutDatum
-            AlonzoEraOnwardsConway
-            BabbageEraOnwardsConway
-            datum
-        )
-        ( case mRefScript of
-            SNothing -> ReferenceScriptNone
-            SJust refScript ->
-              fromShelleyScriptToReferenceScript ShelleyBasedEraConway refScript
-        )
-     where
-      datum = ledgerTxOut ^. L.datumTxOutL
-      mRefScript = ledgerTxOut ^. L.referenceScriptTxOutL
-
-toAlonzoTxOutDatumHash
-  :: TxOutDatum ctx era -> StrictMaybe Plutus.DataHash
-toAlonzoTxOutDatumHash TxOutDatumNone = SNothing
-toAlonzoTxOutDatumHash (TxOutDatumHash _ (ScriptDataHash dh)) = SJust dh
-toAlonzoTxOutDatumHash (TxOutDatumInline{}) = SNothing
-toAlonzoTxOutDatumHash (TxOutSupplementalDatum _ d) =
-  let ScriptDataHash dh = hashScriptDataBytes d
-   in SJust dh
-
-fromAlonzoTxOutDatumHash
-  :: AlonzoEraOnwards era
-  -> StrictMaybe Plutus.DataHash
-  -> TxOutDatum ctx era
-fromAlonzoTxOutDatumHash _ SNothing = TxOutDatumNone
-fromAlonzoTxOutDatumHash w (SJust hash) = TxOutDatumHash w $ ScriptDataHash hash
-
-toBabbageTxOutDatum
-  :: L.Era (ShelleyLedgerEra era)
-  => TxOutDatum ctx era
-  -> Plutus.Datum (ShelleyLedgerEra era)
-toBabbageTxOutDatum TxOutDatumNone = Plutus.NoDatum
-toBabbageTxOutDatum (TxOutDatumHash _ (ScriptDataHash dh)) = Plutus.DatumHash dh
-toBabbageTxOutDatum (TxOutDatumInline _ sd) = scriptDataToInlineDatum sd
-toBabbageTxOutDatum (TxOutSupplementalDatum _ d) =
-  let ScriptDataHash dh = hashScriptDataBytes d
-   in Plutus.DatumHash dh
-
-fromBabbageTxOutDatum
-  :: L.Era ledgerera
-  => AlonzoEraOnwards era
-  -> BabbageEraOnwards era
-  -> Plutus.Datum ledgerera
-  -> TxOutDatum ctx era
-fromBabbageTxOutDatum _ _ Plutus.NoDatum = TxOutDatumNone
-fromBabbageTxOutDatum w _ (Plutus.DatumHash dh) =
-  TxOutDatumHash w $ ScriptDataHash dh
-fromBabbageTxOutDatum _ w (Plutus.Datum binData) =
-  TxOutDatumInline w $ binaryDataToScriptData w binData
 
 -- ----------------------------------------------------------------------------
 -- Transaction input values (era-dependent)
@@ -1142,124 +546,6 @@ deriving instance Eq (TxInsCollateral era)
 
 deriving instance Show (TxInsCollateral era)
 
-data TxInsReference era where
-  TxInsReferenceNone :: TxInsReference era
-  TxInsReference
-    :: BabbageEraOnwards era
-    -> [TxIn]
-    -> TxInsReference era
-
-deriving instance Eq (TxInsReference era)
-
-deriving instance Show (TxInsReference era)
-
--- ----------------------------------------------------------------------------
--- Transaction output values (era-dependent)
---
-
-data TxOutValue era where
-  TxOutValueByron
-    :: L.Coin
-    -> TxOutValue ByronEra
-  TxOutValueShelleyBased
-    :: ( Eq (Ledger.Value (ShelleyLedgerEra era))
-       , Show (Ledger.Value (ShelleyLedgerEra era))
-       )
-    => ShelleyBasedEra era
-    -> L.Value (ShelleyLedgerEra era)
-    -> TxOutValue era
-
-deriving instance Eq (TxOutValue era)
-
-deriving instance Show (TxOutValue era)
-
-instance IsCardanoEra era => ToJSON (TxOutValue era) where
-  toJSON = \case
-    TxOutValueByron ll ->
-      toJSON ll
-    TxOutValueShelleyBased sbe val ->
-      shelleyBasedEraConstraints sbe $ toJSON (fromLedgerValue sbe val)
-
-instance IsShelleyBasedEra era => FromJSON (TxOutValue era) where
-  parseJSON = withObject "TxOutValue" $ \o ->
-    caseShelleyToAllegraOrMaryEraOnwards
-      ( \shelleyToAlleg -> do
-          ll <- o .: "lovelace"
-          let sbe = convert shelleyToAlleg
-          pure $
-            shelleyBasedEraConstraints sbe $
-              TxOutValueShelleyBased sbe $
-                A.mkAdaValue sbe ll
-      )
-      ( \w -> do
-          let l = toList o
-              sbe = convert w
-          vals <- mapM decodeAssetId l
-          pure $
-            shelleyBasedEraConstraints sbe $
-              TxOutValueShelleyBased sbe $
-                toLedgerValue w $
-                  mconcat vals
-      )
-      (shelleyBasedEra @era)
-   where
-    decodeAssetId :: (Aeson.Key, Aeson.Value) -> Aeson.Parser Value
-    decodeAssetId (polid, Aeson.Object assetNameHm) = do
-      let polId = fromString . Text.unpack $ Aeson.toText polid
-      aNameQuantity <- decodeAssets assetNameHm
-      pure . fromList $
-        map (first $ AssetId polId) aNameQuantity
-    decodeAssetId ("lovelace", Aeson.Number sci) =
-      case toBoundedInteger sci of
-        Just (ll :: Word64) ->
-          pure $ fromList [(AdaAssetId, Quantity $ toInteger ll)]
-        Nothing ->
-          fail $ "Expected a Bounded number but got: " <> show sci
-    decodeAssetId wrong = fail $ "Expected a policy id and a JSON object but got: " <> show wrong
-
-    decodeAssets :: Aeson.Object -> Aeson.Parser [(AssetName, Quantity)]
-    decodeAssets assetNameHm =
-      let l = toList assetNameHm
-       in mapM (\(aName, q) -> (,) <$> parseKeyAsAssetName aName <*> decodeQuantity q) l
-
-    parseKeyAsAssetName :: Aeson.Key -> Aeson.Parser AssetName
-    parseKeyAsAssetName aName = runParsecParser parseAssetName (Aeson.toText aName)
-
-    decodeQuantity :: Aeson.Value -> Aeson.Parser Quantity
-    decodeQuantity (Aeson.Number sci) =
-      case toBoundedInteger sci of
-        Just (ll :: Word64) -> return . Quantity $ toInteger ll
-        Nothing -> fail $ "Expected a Bounded number but got: " <> show sci
-    decodeQuantity wrong = fail $ "Expected aeson Number but got: " <> show wrong
-
-lovelaceToTxOutValue
-  :: ()
-  => ShelleyBasedEra era
-  -> L.Coin
-  -> TxOutValue era
-lovelaceToTxOutValue era ll =
-  shelleyBasedEraConstraints era $
-    TxOutValueShelleyBased era $
-      A.mkAdaValue era ll
-
-txOutValueToLovelace :: TxOutValue era -> L.Coin
-txOutValueToLovelace tv =
-  case tv of
-    TxOutValueByron l -> l
-    TxOutValueShelleyBased sbe v -> v ^. A.adaAssetL sbe
-
-txOutValueToValue :: TxOutValue era -> Value
-txOutValueToValue tv =
-  case tv of
-    TxOutValueByron l -> lovelaceToValue l
-    TxOutValueShelleyBased sbe v -> fromLedgerValue sbe v
-
-prettyRenderTxOut :: TxOutInAnyEra -> Text
-prettyRenderTxOut (TxOutInAnyEra _ (TxOut (AddressInEra _ addr) txOutVal _ _)) =
-  serialiseAddress (toAddressAny addr)
-    <> " + "
-    <> renderValue (txOutValueToValue txOutVal)
-
 data TxReturnCollateral ctx era where
   TxReturnCollateralNone
     :: TxReturnCollateral ctx era
@@ -1284,44 +570,16 @@ deriving instance Eq (TxTotalCollateral era)
 
 deriving instance Show (TxTotalCollateral era)
 
--- ----------------------------------------------------------------------------
--- Transaction output datum (era-dependent)
---
-
-data TxOutDatum ctx era where
-  TxOutDatumNone :: TxOutDatum ctx era
-  -- | A transaction output that only specifies the hash of the datum, but
-  -- not the full datum value.
-  TxOutDatumHash
-    :: AlonzoEraOnwards era
-    -> Hash ScriptData
-    -> TxOutDatum ctx era
-  -- | A transaction output that specifies the whole datum value. This can
-  -- only be used in the context of the transaction body (i.e this is a supplemental datum),
-  -- and does not occur in the UTxO. The UTxO only contains the datum hash.
-  TxOutSupplementalDatum
-    :: AlonzoEraOnwards era
-    -> HashableScriptData
-    -> TxOutDatum CtxTx era
-  -- | A transaction output that specifies the whole datum instead of the
-  -- datum hash. Note that the datum map will not be updated with this datum,
-  -- it only exists at the transaction output.
-  TxOutDatumInline
+data TxInsReference era where
+  TxInsReferenceNone :: TxInsReference era
+  TxInsReference
     :: BabbageEraOnwards era
-    -> HashableScriptData
-    -> TxOutDatum ctx era
+    -> [TxIn]
+    -> TxInsReference era
 
-deriving instance Eq (TxOutDatum ctx era)
+deriving instance Eq (TxInsReference era)
 
-deriving instance Show (TxOutDatum ctx era)
-
-{-# COMPLETE TxOutDatumNone, TxOutDatumHash, TxOutSupplementalDatum, TxOutDatumInline #-}
-
-parseHash :: SerialiseAsRawBytes (Hash a) => Parsec.Parser (Hash a)
-parseHash = do
-  str <- some Parsec.hexDigit <?> "hash"
-  failEitherWith (\e -> "Failed to parse hash: " ++ displayError e) $
-    deserialiseFromRawBytesHex (BSC.pack str)
+deriving instance Show (TxInsReference era)
 
 -- ----------------------------------------------------------------------------
 -- Transaction fees
@@ -2064,8 +1322,7 @@ data TxBodyError
   | TxBodyEmptyTxIns
   | TxBodyEmptyTxInsCollateral
   | TxBodyEmptyTxOuts
-  | TxBodyOutputNegative !Quantity !TxOutInAnyEra
-  | TxBodyOutputOverflow !Quantity !TxOutInAnyEra
+  | TxBodyOutputError !TxOutputError
   | TxBodyMetadataError ![(Word64, TxMetadataRangeError)]
   | TxBodyInIxOverflow !TxIn
   | TxBodyMissingProtocolParams
@@ -2085,16 +1342,7 @@ instance Error TxBodyError where
       "Transaction body has no collateral inputs, but uses Plutus scripts"
     TxBodyEmptyTxOuts ->
       "Transaction body has no outputs"
-    TxBodyOutputNegative (Quantity q) txout ->
-      "Negative quantity ("
-        <> pretty q
-        <> ") in transaction output: "
-        <> pretty txout
-    TxBodyOutputOverflow (Quantity q) txout ->
-      "Quantity too large ("
-        <> pretty q
-        <> " >= 2^64) in transaction output: "
-        <> pretty txout
+    TxBodyOutputError err -> prettyError err
     TxBodyMetadataError [(k, err)] ->
       "Error in metadata entry " <> pretty k <> ": " <> prettyError err
     TxBodyMetadataError errs ->
@@ -2279,35 +1527,41 @@ validateTxBodyContent
           ShelleyBasedEraShelley -> do
             validateTxIns txIns
             guardShelleyTxInsOverflow (map fst txIns)
-            validateTxOuts sbe txOuts
+            first TxBodyOutputError $
+              validateTxOuts sbe txOuts
             validateMetadata txMetadata
           ShelleyBasedEraAllegra -> do
             validateTxIns txIns
             guardShelleyTxInsOverflow (map fst txIns)
-            validateTxOuts sbe txOuts
+            first TxBodyOutputError $
+              validateTxOuts sbe txOuts
             validateMetadata txMetadata
           ShelleyBasedEraMary -> do
             validateTxIns txIns
             guardShelleyTxInsOverflow (map fst txIns)
-            validateTxOuts sbe txOuts
+            first TxBodyOutputError $
+              validateTxOuts sbe txOuts
             validateMetadata txMetadata
           ShelleyBasedEraAlonzo -> do
             validateTxIns txIns
             guardShelleyTxInsOverflow (map fst txIns)
-            validateTxOuts sbe txOuts
+            first TxBodyOutputError $
+              validateTxOuts sbe txOuts
             validateMetadata txMetadata
             validateTxInsCollateral txInsCollateral languages
             validateProtocolParameters txProtocolParams languages
           ShelleyBasedEraBabbage -> do
             validateTxIns txIns
             guardShelleyTxInsOverflow (map fst txIns)
-            validateTxOuts sbe txOuts
+            first TxBodyOutputError $
+              validateTxOuts sbe txOuts
             validateMetadata txMetadata
             validateTxInsCollateral txInsCollateral languages
             validateProtocolParameters txProtocolParams languages
           ShelleyBasedEraConway -> do
             validateTxIns txIns
-            validateTxOuts sbe txOuts
+            first TxBodyOutputError $
+              validateTxOuts sbe txOuts
             validateMetadata txMetadata
             validateTxInsCollateral txInsCollateral languages
             validateProtocolParameters txProtocolParams languages
@@ -2346,52 +1600,16 @@ validateTxInsCollateral txInsCollateral languages =
     TxInsCollateral _ collateralTxIns ->
       guardShelleyTxInsOverflow collateralTxIns
 
-validateTxOuts :: ShelleyBasedEra era -> [TxOut CtxTx era] -> Either TxBodyError ()
-validateTxOuts sbe txOuts = do
-  let era = toCardanoEra sbe
-  cardanoEraConstraints era $
-    sequence_
-      [ do
-          positiveOutput era (txOutValueToValue v) txout
-          outputDoesNotExceedMax era (txOutValueToValue v) txout
-      | txout@(TxOut _ v _ _) <- txOuts
-      ]
-
 inputIndexDoesNotExceedMax :: [(TxIn, a)] -> Either TxBodyError ()
 inputIndexDoesNotExceedMax txIns =
   for_ txIns $ \(txin@(TxIn _ (TxIx txix)), _) ->
     guard (fromIntegral txix <= maxShelleyTxInIx) ?! TxBodyInIxOverflow txin
-
-outputDoesNotExceedMax
-  :: ()
-  => CardanoEra era
-  -> Value
-  -> TxOut CtxTx era
-  -> Either TxBodyError ()
-outputDoesNotExceedMax era v txout =
-  case [q | (_, q) <- toList v, q > maxTxOut] of
-    [] -> Right ()
-    q : _ -> Left (TxBodyOutputOverflow q (txOutInAnyEra era txout))
-
-positiveOutput
-  :: ()
-  => CardanoEra era
-  -> Value
-  -> TxOut CtxTx era
-  -> Either TxBodyError ()
-positiveOutput era v txout =
-  case [q | (_, q) <- toList v, q < 0] of
-    [] -> Right ()
-    q : _ -> Left (TxBodyOutputNegative q (txOutInAnyEra era txout))
 
 txBodyContentHasTxIns :: TxIns BuildTx era -> Either TxBodyError ()
 txBodyContentHasTxIns txIns = guard (not (null txIns)) ?! TxBodyEmptyTxIns
 
 maxShelleyTxInIx :: Word
 maxShelleyTxInIx = fromIntegral $ maxBound @Word16
-
-maxTxOut :: Quantity
-maxTxOut = fromIntegral (maxBound :: Word64)
 
 {-# DEPRECATED createAndValidateTransactionBody "Use createTransactionBody instead" #-}
 createAndValidateTransactionBody
@@ -2530,117 +1748,6 @@ fromLedgerTxInsReference sbe txBody =
     (const TxInsReferenceNone)
     (\w -> TxInsReference w $ map fromShelleyTxIn . toList $ txBody ^. L.referenceInputsTxBodyL)
     sbe
-
-fromLedgerTxOuts
-  :: forall era
-   . ShelleyBasedEra era
-  -> Ledger.TxBody (ShelleyLedgerEra era)
-  -> TxBodyScriptData era
-  -> [TxOut CtxTx era]
-fromLedgerTxOuts sbe body scriptdata =
-  case sbe of
-    ShelleyBasedEraShelley ->
-      [fromShelleyTxOut sbe txout | txout <- toList (body ^. L.outputsTxBodyL)]
-    ShelleyBasedEraAllegra ->
-      [fromShelleyTxOut sbe txout | txout <- toList (body ^. L.outputsTxBodyL)]
-    ShelleyBasedEraMary ->
-      [fromShelleyTxOut sbe txout | txout <- toList (body ^. L.outputsTxBodyL)]
-    ShelleyBasedEraAlonzo ->
-      [ fromAlonzoTxOut
-          AlonzoEraOnwardsAlonzo
-          txdatums
-          txout
-      | let txdatums = selectTxDatums scriptdata
-      , txout <- toList (body ^. L.outputsTxBodyL)
-      ]
-    ShelleyBasedEraBabbage ->
-      [ fromBabbageTxOut
-          BabbageEraOnwardsBabbage
-          txdatums
-          txouts
-      | let txdatums = selectTxDatums scriptdata
-      , txouts <- toList (body ^. L.outputsTxBodyL)
-      ]
-    ShelleyBasedEraConway ->
-      [ fromBabbageTxOut
-          BabbageEraOnwardsConway
-          txdatums
-          txouts
-      | let txdatums = selectTxDatums scriptdata
-      , txouts <- toList (body ^. L.outputsTxBodyL)
-      ]
-
-selectTxDatums
-  :: TxBodyScriptData era
-  -> Map L.DataHash (L.Data (ShelleyLedgerEra era))
-selectTxDatums TxBodyNoScriptData = Map.empty
-selectTxDatums (TxBodyScriptData _ (Alonzo.TxDats' datums) _) = datums
-
-fromAlonzoTxOut
-  :: forall era ledgerera
-   . AlonzoEraOnwards era
-  -> Map L.DataHash (L.Data ledgerera)
-  -> L.TxOut (ShelleyLedgerEra era)
-  -> TxOut CtxTx era
-fromAlonzoTxOut w txdatums txOut =
-  alonzoEraOnwardsConstraints w $
-    TxOut
-      (fromShelleyAddr shelleyBasedEra (txOut ^. L.addrTxOutL))
-      (TxOutValueShelleyBased sbe (txOut ^. L.valueTxOutL))
-      (fromAlonzoTxOutDatum w txdatums (txOut ^. L.dataHashTxOutL))
-      ReferenceScriptNone
- where
-  sbe :: ShelleyBasedEra era = convert w
-
-fromAlonzoTxOutDatum
-  :: ()
-  => AlonzoEraOnwards era
-  -> Map L.DataHash (L.Data ledgerera)
-  -> StrictMaybe L.DataHash
-  -> TxOutDatum CtxTx era
-fromAlonzoTxOutDatum w txdatums = \case
-  SNothing -> TxOutDatumNone
-  SJust dh
-    | Just d <- Map.lookup dh txdatums ->
-        TxOutSupplementalDatum w (fromAlonzoData d)
-    | otherwise -> TxOutDatumHash w (ScriptDataHash dh)
-
-fromBabbageTxOut
-  :: forall era
-   . BabbageEraOnwards era
-  -> Map L.DataHash (L.Data (ShelleyLedgerEra era))
-  -> L.TxOut (ShelleyLedgerEra era)
-  -> TxOut CtxTx era
-fromBabbageTxOut w txdatums txout =
-  babbageEraOnwardsConstraints w $
-    TxOut
-      (fromShelleyAddr shelleyBasedEra (txout ^. L.addrTxOutL))
-      (TxOutValueShelleyBased sbe (txout ^. L.valueTxOutL))
-      babbageTxOutDatum
-      ( case txout ^. L.referenceScriptTxOutL of
-          SNothing -> ReferenceScriptNone
-          SJust rScript -> fromShelleyScriptToReferenceScript shelleyBasedEra rScript
-      )
- where
-  sbe :: ShelleyBasedEra era = convert w
-
-  -- NOTE: This is different to 'fromBabbageTxOutDatum' as it may resolve
-  -- 'DatumHash' values using the datums included in the transaction.
-  babbageTxOutDatum :: TxOutDatum CtxTx era
-  babbageTxOutDatum =
-    babbageEraOnwardsConstraints w $
-      case txout ^. L.datumTxOutL of
-        L.NoDatum -> TxOutDatumNone
-        L.DatumHash dh -> resolveDatumInTx dh
-        L.Datum d -> TxOutDatumInline w $ binaryDataToScriptData w d
-
-  resolveDatumInTx :: L.DataHash -> TxOutDatum CtxTx era
-  resolveDatumInTx dh
-    | Just d <- Map.lookup dh txdatums =
-        TxOutSupplementalDatum
-          (convert w)
-          (fromAlonzoData d)
-    | otherwise = TxOutDatumHash (convert w) (ScriptDataHash dh)
 
 fromLedgerTxTotalCollateral
   :: ShelleyBasedEra era
@@ -2835,7 +1942,7 @@ makeByronTransactionBody txIns txOuts = do
   outs' <- NonEmpty.nonEmpty txOuts ?! TxBodyEmptyTxOuts
   outs'' <-
     traverse
-      (\out -> toByronTxOut out ?! classifyRangeError out)
+      (first TxBodyOutputError . toByronTxOut)
       outs'
   return $
     CBOR.reAnnotate CBOR.byronProtVer $
@@ -2845,15 +1952,6 @@ makeByronTransactionBody txIns txOuts = do
  where
   maxByronTxInIx :: Word
   maxByronTxInIx = fromIntegral (maxBound :: Word32)
-
-classifyRangeError :: TxOut CtxTx ByronEra -> TxBodyError
-classifyRangeError txout =
-  case txout of
-    TxOut (AddressInEra ByronAddressInAnyEra ByronAddress{}) (TxOutValueByron value) _ _
-      | value < 0 -> TxBodyOutputNegative (lovelaceToQuantity value) (txOutInAnyEra ByronEra txout)
-      | otherwise -> TxBodyOutputOverflow (lovelaceToQuantity value) (txOutInAnyEra ByronEra txout)
-    TxOut (AddressInEra ByronAddressInAnyEra (ByronAddress _)) (TxOutValueShelleyBased w _) _ _ -> case w of {}
-    TxOut (AddressInEra (ShelleyAddressInEra sbe) ShelleyAddress{}) _ _ _ -> case sbe of {}
 
 convTxIns :: TxIns BuildTx era -> Set L.TxIn
 convTxIns txIns = fromList (map (toShelleyTxIn . fst) txIns)
@@ -2878,15 +1976,6 @@ convTotalCollateral txTotalCollateral =
   case txTotalCollateral of
     TxTotalCollateralNone -> SNothing
     TxTotalCollateral _ totCollLovelace -> SJust totCollLovelace
-
-convTxOuts
-  :: forall ctx era ledgerera
-   . HasCallStack
-  => ShelleyLedgerEra era ~ ledgerera
-  => ShelleyBasedEra era
-  -> [TxOut ctx era]
-  -> Seq.StrictSeq (Ledger.TxOut ledgerera)
-convTxOuts sbe txOuts = fromList $ map (toShelleyTxOutAny sbe) txOuts
 
 convCertificates
   :: ShelleyBasedEra era
@@ -3643,40 +2732,6 @@ makeShelleyTransactionBody
     txAuxData :: Maybe (L.TxAuxData E.ConwayEra)
     txAuxData = toAuxiliaryData sbe txMetadata txAuxScripts
 
--- | A variant of 'toShelleyTxOutAny that is used only internally to this module
--- that works with a 'TxOut' in any context (including CtxTx) by ignoring
--- embedded datums (taking only their hash).
-toShelleyTxOutAny
-  :: forall ctx era ledgerera
-   . HasCallStack
-  => ShelleyLedgerEra era ~ ledgerera
-  => ShelleyBasedEra era
-  -> TxOut ctx era
-  -> Ledger.TxOut ledgerera
-toShelleyTxOutAny sbe = shelleyBasedEraConstraints sbe $ \case
-  TxOut addr (TxOutValueShelleyBased _ value) txoutdata refScript ->
-    caseShelleyToMaryOrAlonzoEraOnwards
-      (const $ L.mkBasicTxOut (toShelleyAddr addr) value)
-      ( \case
-          AlonzoEraOnwardsAlonzo ->
-            L.mkBasicTxOut (toShelleyAddr addr) value
-              & L.dataHashTxOutL
-                .~ toAlonzoTxOutDatumHash txoutdata
-          AlonzoEraOnwardsBabbage ->
-            L.mkBasicTxOut (toShelleyAddr addr) value
-              & L.datumTxOutL
-                .~ toBabbageTxOutDatum txoutdata
-              & L.referenceScriptTxOutL
-                .~ refScriptToShelleyScript sbe refScript
-          AlonzoEraOnwardsConway ->
-            L.mkBasicTxOut (toShelleyAddr addr) value
-              & L.datumTxOutL
-                .~ toBabbageTxOutDatum txoutdata
-              & L.referenceScriptTxOutL
-                .~ refScriptToShelleyScript sbe refScript
-      )
-      sbe
-
 -- ----------------------------------------------------------------------------
 -- Script witnesses within the tx body
 --
@@ -3927,18 +2982,6 @@ collectTxBodyScriptWitnesses
         | (ix, _, witness) <- indexTxProposalProcedures txp
         ]
 
-getSupplementalDatums
-  :: AlonzoEraOnwards era -> [TxOut CtxTx era] -> L.TxDats (ShelleyLedgerEra era)
-getSupplementalDatums eon [] = alonzoEraOnwardsConstraints eon mempty
-getSupplementalDatums eon txouts =
-  alonzoEraOnwardsConstraints eon $
-    L.TxDats $
-      fromList
-        [ (L.hashData ledgerData, ledgerData)
-        | TxOut _ _ (TxOutSupplementalDatum _ d) _ <- txouts
-        , let ledgerData = toAlonzoData d
-        ]
-
 collectTxBodyScriptWitnessRequirements
   :: forall era
    . IsShelleyBasedEra era
@@ -4007,6 +3050,20 @@ collectTxBodyScriptWitnessRequirements
             , txVotingWits
             , txProposalWits
             ]
+
+getSupplementalDatums
+  :: AlonzoEraOnwards era
+  -> [TxOut CtxTx era]
+  -> L.TxDats (ShelleyLedgerEra era)
+getSupplementalDatums eon [] = alonzoEraOnwardsConstraints eon mempty
+getSupplementalDatums eon txouts =
+  alonzoEraOnwardsConstraints eon $
+    L.TxDats $
+      fromList
+        [ (L.hashData ledgerData, ledgerData)
+        | TxOut _ _ (TxOutSupplementalDatum _ d) _ <- txouts
+        , let ledgerData = toAlonzoData d
+        ]
 
 extractWitnessableTxIns
   :: AlonzoEraOnwards era
@@ -4191,23 +3248,3 @@ getReferenceInputsSizeForTxIds beo utxo txIds = babbageEraOnwardsConstraints beo
 calculateExecutionUnitsLovelace :: Ledger.Prices -> ExecutionUnits -> Maybe L.Coin
 calculateExecutionUnitsLovelace prices eUnits =
   return $ Alonzo.txscriptfee prices (toAlonzoExUnits eUnits)
-
--- ----------------------------------------------------------------------------
--- Inline data
---
-
--- | Conversion of ScriptData to binary data which allows for the storage of data
--- onchain within a transaction output.
-scriptDataToInlineDatum :: L.Era ledgerera => HashableScriptData -> L.Datum ledgerera
-scriptDataToInlineDatum d =
-  L.Datum . L.dataToBinaryData $ toAlonzoData d
-
-binaryDataToScriptData
-  :: L.Era ledgerera
-  => BabbageEraOnwards era
-  -> L.BinaryData ledgerera
-  -> HashableScriptData
-binaryDataToScriptData BabbageEraOnwardsBabbage d =
-  fromAlonzoData $ L.binaryDataToData d
-binaryDataToScriptData BabbageEraOnwardsConway d =
-  fromAlonzoData $ L.binaryDataToData d
