+++ cardano-api/test/cardano-api-golden/Test/Golden/ErrorsSpec.hs
@@ -415,8 +415,8 @@ test_TxBodyError =
     [ ("TxBodyEmptyTxIns", TxBodyEmptyTxIns)
     , ("TxBodyEmptyTxInsCollateral", TxBodyEmptyTxInsCollateral)
     , ("TxBodyEmptyTxOuts", TxBodyEmptyTxOuts)
-    , ("TxBodyOutputNegative", TxBodyOutputNegative 1 txOutInAnyEra1)
-    , ("TxBodyOutputOverflow", TxBodyOutputOverflow 1 txOutInAnyEra1)
+    , ("TxBodyOutputNegative", TxBodyOutputError $ TxOutputNegative 1 txOutInAnyEra1)
+    , ("TxBodyOutputOverflow", TxBodyOutputError $ TxOutputOverflow 1 txOutInAnyEra1)
     , ("TxBodyMetadataError", TxBodyMetadataError [(1, TxMetadataBytesTooLong 2)])
     , ("TxBodyMissingProtocolParams", TxBodyMissingProtocolParams)
     , ("TxBodyInIxOverflow", TxBodyInIxOverflow txin1)
